+++ src/web/spa-navigation.test.ts
@@ -109,6 +109,7 @@ function makeState(
     writeContextFile: () => {},
     updateAgentAvatar: () => {},
     setAgentEnabled: () => true,
+    setAgentModel: () => true,
     resolveChatImage: async () => null,
     resolveDiscordGuildImage: async () => null,
     ...overrides,