+++ src/web/system.test.ts
@@ -98,6 +98,7 @@ function makeState(
     writeContextFile: () => {},
     updateAgentAvatar: () => {},
     setAgentEnabled: () => true,
+    setAgentModel: () => true,
   };
 }